Move the Account class to the common package.
<h1>Rationale:</h1>

This commit provides no functional changes.  It simply moves the Account class to the common/model package to reflect that it is tied to both chat and experiences.
