+++ app/src/main/java/com/pajato/android/gamechat/chat/adapter/SelectableMemberItem.java
@@ -17,7 +17,7 @@
 
 package com.pajato.android.gamechat.chat.adapter;
 
-import com.pajato.android.gamechat.chat.model.Account;
+import com.pajato.android.gamechat.common.model.Account;
 import com.pajato.android.gamechat.chat.model.Group;
 import com.pajato.android.gamechat.database.GroupManager;
