+++ app/src/main/java/com/pajato/android/gamechat/chat/fragment/CreateGroupFragment.java
@@ -23,7 +23,7 @@ import android.widget.RadioGroup;
 
 import com.pajato.android.gamechat.R;
 import com.pajato.android.gamechat.chat.BaseCreateFragment;
-import com.pajato.android.gamechat.chat.model.Account;
+import com.pajato.android.gamechat.common.model.Account;
 import com.pajato.android.gamechat.chat.model.Group;
 import com.pajato.android.gamechat.chat.model.Room;
 import com.pajato.android.gamechat.database.AccountManager;
