+++ app/src/main/java/com/pajato/android/gamechat/common/InvitationManager.java
@@ -19,7 +19,7 @@ package com.pajato.android.gamechat.common;
 
 import android.support.annotation.NonNull;
 
-import com.pajato.android.gamechat.chat.model.Account;
+import com.pajato.android.gamechat.common.model.Account;
 import com.pajato.android.gamechat.chat.model.Group;
 import com.pajato.android.gamechat.database.AccountManager;
 import com.pajato.android.gamechat.database.DBUtils;
