+++ app/src/main/java/com/pajato/android/gamechat/database/handler/MemberChangeHandler.java
@@ -23,7 +23,7 @@ import android.util.Log;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.ValueEventListener;
-import com.pajato.android.gamechat.chat.model.Account;
+import com.pajato.android.gamechat.common.model.Account;
 import com.pajato.android.gamechat.event.AppEventManager;
 import com.pajato.android.gamechat.event.MemberChangeEvent;
