+++ app/src/main/java/com/pajato/android/gamechat/event/AuthenticationChangeHandled.java
@@ -17,7 +17,7 @@
 
 package com.pajato.android.gamechat.event;
 
-import com.pajato.android.gamechat.chat.model.Account;
+import com.pajato.android.gamechat.common.model.Account;
 
 /**
  * Provides an event type indicating that an account state change has just finished being handled.
